refactor(packages): Merge non-client packages into common `odg-core-libs`
