+++ .github/workflows/build.yaml
@@ -60,19 +60,15 @@ jobs:
           version=${{ needs.setuptools-version.outputs.setuptools-version }}
           echo "version: ${version}"
 
-          # pass finalised version to setups
-          export ODG_SERVICE_PACKAGE_VERSION=${version}
-          export ODG_EXTENSIONS_PACKAGE_VERSION=${version}
-          export ODG_UTILS_PACKAGE_VERSION=${version}
+          # pass finalised version to setup
+          export ODG_CORE_LIBS_VERSION=${version}
 
           pkg_dir=dist
           mkdir -p ${pkg_dir}
           pkg_dir="$(readlink -f dist)"
 
           for path in \
-            setup.utils.py \
-            setup.service.py \
-            setup.extensions.py \
+            setup.py \
           ; do
           echo "building distribution package from ${path}"
           python3 ${path} \
@@ -89,9 +85,7 @@ jobs:
           resources_file=resources.yaml
 
           for package in \
-            delivery-gear-utils \
-            ocm-gear-extensions \
-            ocm-gear-service \
+            odg-core-libs \
             ; do
             path="$(echo $package | tr - _)-*"
             outf="${pkg_dir}/${package}.tar.gz"
@@ -179,9 +173,9 @@ jobs:
           echo "installing linters"
           if ! pip3 install --pre --upgrade --break-system-packages \
             --find-links /tmp/dist \
-            -r requirements.extensions.txt \
-            -r requirements.service.txt \
-            -r requirements.utils.txt \
+            -r requirements.bdba-client.txt \
+            -r requirements.odg-client.txt \
+            -r requirements.txt \
             numpy \
             scipy \
             bandit \
