+++ Dockerfile.extensions.crypto
@@ -23,7 +23,7 @@ RUN --mount=type=bind,source=/dist,target=/dist \
     python3-dev \
     syft \
 && CFLAGS='-Wno-int-conversion' \
-    pip3 install --upgrade --no-cache-dir --find-links ./dist ocm-gear-extensions \
+    pip3 install --upgrade --no-cache-dir --find-links ./dist odg-core-libs \
 && apk del --no-cache \
     libc-dev \
     libffi-dev \
