+++ Dockerfile.extensions.db-manager
@@ -10,7 +10,7 @@ RUN --mount=type=bind,source=/dist,target=/dist \
     postgresql16-client \
     python3-dev \
 && CFLAGS='-Wno-int-conversion' \
-    pip3 install --upgrade --no-cache-dir --find-links ./dist ocm-gear-service \
+    pip3 install --upgrade --no-cache-dir --find-links ./dist odg-core-libs \
 && apk del --no-cache \
     libc-dev \
     libffi-dev \
