+++ README.md
@@ -14,7 +14,7 @@ cluster.
 
 # Local Development
 
-Delivery-Service and Extensions require a Python runtime environment (see `setup*.py` for details) to
+Delivery-Service and Extensions require a Python runtime environment (see `setup.py` for details) to
 run. Typically, the Python3 version from greatest released version of
 [Alpine](https://endoflife.date/alpine) Linux is used/tested (see `Dockerfile.*`). Greater or smaller
 versions _may_ work, but are typically untested.
