+++ requirements-dev.txt
@@ -1,8 +1,8 @@
--r requirements.extensions.txt
--r requirements.service.txt
--r requirements.utils.txt
+-r requirements.bdba-client.txt
+-r requirements.odg-client.txt
+-r requirements.txt
 aiohttp-devtools
 aiosqlite
+greenlet
 numpy
 scipy
-greenlet
