--- requirements.utils.txt
@@ -1,12 +1,0 @@
-aiohttp
-bdba
-boto3
-cachetools
-dacite
-gardener-cicd-libs
-github3.py
-kubernetes
-requests
-urllib3
-watchdog
-python-dateutil
